Move string utility methods in FindCommandParser to StringUtil
